requirements.txt
langgraph==0.2.0
langchain==0.2.0
langchain-openai==0.2.0
python-dotenv==1.0.0
pydantic==2.8.0
fastapi==0.115.0
uvicorn==0.32.0
